pyproject.toml
[tool.poetry]
name = "smart_meter_to_openhab"
version = "0.5.2"
description = "Pushing data of ISKRA MT175 smart meter to openhab"
authors = ["Heiko Bauer <heiko_bauer@icloud.com>"]
repository = "https://github.com/die-bauerei/smart-meter-to-openhab"
readme = "README.md"
packages = [
    {include = "smart_meter_to_openhab"},
    {include = "smart_meter_to_openhab_scripts"}
]

[tool.poetry.scripts]
smart_meter_to_openhab = "smart_meter_to_openhab_scripts.main:main"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pydantic = "^2.4.1"
mypy = "^1.5.1"
types-requests = "^2.31.0.10"
types-pyserial = "^3.5.0.10"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
